% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocking-evaluations.R
\name{block.ids.from.blocking}
\alias{block.ids.from.blocking}
\title{Returns the block ids associated with a blocking method.}
\usage{
block.ids.from.blocking(blocking)
}
\arguments{
\item{blocking}{A list of the blocks.}
}
\value{
A list of the blocks ids that corresponds to each block
}
\description{
Returns the block ids associated with a blocking method.
}
\examples{
tlsh.blocks <- block_setup_v2(r.set = RLdata500[1:250,c(-2,-4)], b=10, save_signature=FALSE, k=1)
block.ids.from.blocking(tlsh.blocks)
}
