% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlsh.R
\name{block_setup_v2}
\alias{block_setup_v2}
\title{Function that divides all records into bins using locality sensitive hashing and using TLSH (based upon community detection technique)}
\usage{
block_setup_v2(r.set, b = 22, save_signature = FALSE, k = 5)
}
\arguments{
\item{r.set}{Record set (shingled records)}

\item{b}{Band}

\item{save_signature}{Flag of whether or not to save the signature}

\item{k}{Shingle size}
}
\value{
List of blocks where a particular index is the record id in the original
data set
}
\description{
import blink
}
\examples{
r.set <- RLdata500[1:3,c(-2)]
block_setup_v2(r.set = RLdata500[1:3,c(-2)], b=22, save_signature=FALSE, k=2)
}
