% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlsh.R
\name{eval.blocksetup}
\alias{eval.blocksetup}
\title{Function to evaluate the blocking step}
\usage{
eval.blocksetup(dat, k = 5, b = 21, key)
}
\arguments{
\item{dat}{Data set}

\item{k}{Parameter k, which is the number of shingle, tokens, or grams to break the string into}

\item{b}{Number of buckets}

\item{key}{Unique identifier}
}
\value{
Recall and runtime
}
\description{
import blink
}
\examples{
r.set <- RLdata500[1:50,c(-2)]
eval.blocksetup(r.set, k=2, b=22, key=identity.RLdata500)
}
