% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhash_v2.R
\name{my_hash}
\alias{my_hash}
\title{Function that applies a hash function to each column of the band from the
signature matrix
import bit64}
\usage{
my_hash(a_band)
}
\arguments{
\item{a_band}{Band from the signature matrix M}
}
\value{
a 64 bit integer
}
\description{
Function that applies a hash function to each column of the band from the
signature matrix
import bit64
}
\examples{
band1 <- c(2,1,2,1,2)
band2 <- c(4,5,2,1,9)
combined_band <- rbind(band1,band2)
my_hash(combined_band)
}
