% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhash_v2.R
\name{primest}
\alias{primest}
\title{Function to generate all primes larger than an integer n1 (lower limit) and less than any other integer n2 (upper limit)}
\usage{
primest(n1 = 1, n2)
}
\arguments{
\item{n1}{An integer taken to be 1 as the default}

\item{n2}{Any integer n2}
}
\value{
Generates all prime numbers with the above constraints
}
\description{
Function to generate all primes larger than an integer n1 (lower limit) and less than any other integer n2 (upper limit)
}
\examples{
primest(1, 5)
primest(1, 17)
}
