% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocking-evaluations.R
\name{reduction.ratio.from.blocking}
\alias{reduction.ratio.from.blocking}
\title{Returns the reduction ratio associated with a blocking method}
\usage{
reduction.ratio.from.blocking(blocking)
}
\arguments{
\item{blocking}{The actual blocks}
}
\value{
The reduction ratio
}
\description{
Returns the reduction ratio associated with a blocking method
}
\examples{
tlsh.blocks <- block_setup_v2(r.set = RLdata500[1:50,c(-2,-4)], b=10, save_signature=FALSE, k=1)
reduction.ratio.from.blocking(tlsh.blocks)
}
