% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocking-evaluations.R
\name{reduction.ratio}
\alias{reduction.ratio}
\title{Returns the reduction ratio associated with a blocking method}
\usage{
reduction.ratio(block.labels)
}
\arguments{
\item{block.labels}{A list of the blocks labels.}
}
\value{
The reduction ratio
}
\description{
Returns the reduction ratio associated with a blocking method
}
\examples{
tlsh.blocks <- block_setup_v2(r.set = RLdata500[1:50,c(-2)], b=22, save_signature=FALSE, k=2)
block.ids <- block.ids.from.blocking(tlsh.blocks)
reduction.ratio(block.ids)
}
