% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhash_v2.R
\name{rhash_funcs}
\alias{rhash_funcs}
\title{Function to generate a vector of random hash functions (or optionally one vector-valued function)}
\usage{
rhash_funcs(n, size, vector.valued, perfect = FALSE)
}
\arguments{
\item{n}{Number of random hash functions}

\item{size}{Range of each size}

\item{vector.valued}{Flag for outputing vector of functions or vector-valued function}

\item{perfect}{Flag for whether a perfect permutation should be done, or just a hash function}
}
\value{
Vector of n hash functions or a function which will take a number and return a vector of n different hashes of it
}
\description{
Function to generate a vector of random hash functions (or optionally one vector-valued function)
}
\examples{
rhash_funcs(1, 1, vector.valued=FALSE, perfect=FALSE)
rhash_funcs(5, 1, vector.valued=FALSE, perfect=FALSE)
}
