% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minhash_v2.R
\name{shingles}
\alias{shingles}
\title{Function to shingle (token or gram) a string into its k components}
\usage{
shingles(record, k)
}
\arguments{
\item{record}{String or record}

\item{k}{Parameter k, which is the number of shingle, tokens, or grams to break the string into}
}
\value{
Computes the shingled (tokened or grammed) version of a string
}
\description{
Function to shingle (token or gram) a string into its k components
}
\examples{
shingles("Alexander",2)
shingles("Alexander Smith", 2)
}
