% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlsh-package.R
\name{tlsh-exports}
\alias{tlsh-exports}
\alias{\%>\%}
\title{tlsh exported operators}
\description{
The following functions are imported and then re-exported
from the tlsh package to enable use of the magrittr
pipe operator with no additional library calls
}
