% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{tlsh_dist}
\alias{tlsh_dist}
\title{Compute distance between two TLSH objects}
\usage{
tlsh_dist(x, y)
}
\arguments{
\item{x, y}{TLSH hash objects}
}
\value{
distance value
}
\description{
Compute distance between two TLSH objects
}
