% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{tlsh_is_valid}
\alias{tlsh_is_valid}
\title{Test if a \code{tlsh} hash object is valid}
\usage{
tlsh_is_valid(x)
}
\arguments{
\item{x}{a \code{tlsh} object}
}
\description{
Test if a \code{tlsh} hash object is valid
}
