% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{tlsh_reset}
\alias{tlsh_reset}
\title{Clear content and hash computation from a \code{tlsh} object}
\usage{
tlsh_reset(x)
}
\arguments{
\item{x}{a \code{tlsh} object}
}
\description{
Modestly lower-cost function that creating a new object.
}
\examples{
doc1 <- as.character(xml2::read_html(system.file("extdat", "index.html", package="tlsh")))

tlsh() \%>\%
  tlsh_update(doc1) \%>\%
  tlsh_finalize() -> x

tlsh_hash(x)

x <- tlsh_reset(x)
}
