% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlsh-hash.R
\name{tlsh_simple_diff}
\alias{tlsh_simple_diff}
\title{Compute the difference between two character hashes}
\usage{
tlsh_simple_diff(x, y)
}
\arguments{
\item{x, y}{two hash fingerprints to compare}
}
\value{
\code{NA} will be returned if \code{x} or \code{y} are not valid hashes
}
\description{
Compute the difference between two character hashes
}
