% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlsh-hash.R
\name{tlsh_simple_hash}
\alias{tlsh_simple_hash}
\alias{tlsh_simple_diff}
\title{Compute TLSH hash for a character or raw vector and return hash fingerprint}
\usage{
tlsh_simple_hash(x)

tlsh_simple_diff(x, y)
}
\arguments{
\item{x}{length 1 \code{character} or \code{raw} vector}

\item{x, y}{two hash fingerprints to compare}
}
\description{
Compute TLSH hash for a character or raw vector and return hash fingerprint
}
