% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlsh-package.R
\docType{package}
\name{tlsh}
\alias{tlsh}
\alias{tlsh-package}
\title{Local Sensitivity Hashing Using the 'Trend Micro' 'TLSH' Implementation}
\description{
'Trend Micro' provides an open source library \url{https://github.com/trendmicro/tlsh/}
for local sensitivity hashing. Methods are provided to compute and compare
hashes from character/byte streams.
}
\author{
Bob Rudis (bob@rud.is)
}
