% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{tlsh_finalize}
\alias{tlsh_finalize}
\title{Finalize a \code{tlsh} object hash}
\usage{
tlsh_finalize(x, force = FALSE)
}
\arguments{
\item{x}{a \code{tlsh} object}

\item{force}{if \code{TRUE} an attempt will be made to forego minimum content-length
and content-body requirements when computing the hash. \strong{Not recommended}.
Default: \code{FALSE}}
}
\value{
\code{tlsh} object
}
\description{
Calling this "finalizes" the \code{tlsh} object, which means the actual hash value
for all the content added to it is computed. This should be the the last call
after all _update_s.
}
\examples{
doc1 <- as.character(xml2::read_html(system.file("extdat", "index.html", package="tlsh")))

tlsh() \%>\%
  tlsh_update(doc1) \%>\%
  tlsh_finalize() -> x

tlsh_hash(x)

x
}
