% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{tlsh_hash}
\alias{tlsh_hash}
\title{Retrieve the hex-encoded hash string for a \code{tlsh} object}
\usage{
tlsh_hash(x)
}
\arguments{
\item{x}{a \code{tlsh} object}
}
\value{
length 1 character vector
}
\description{
Retrieve the hex-encoded hash string for a \code{tlsh} object
}
\examples{
doc1 <- as.character(xml2::read_html(system.file("extdat", "index.html", package="tlsh")))

tlsh() \%>\%
  tlsh_update(doc1) \%>\%
  tlsh_finalize() -> x

tlsh_hash(x)
}
