% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{tlsh_reset}
\alias{tlsh_reset}
\title{Clear content and hash computation from a \code{tlsh} object}
\usage{
tlsh_reset(x)
}
\arguments{
\item{x}{a \code{tlsh} object}
}
\description{
Modestly lower-cost function that creating a new object.
}
