% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlsh-hash.R
\name{tlsh_simple_hash}
\alias{tlsh_simple_hash}
\title{Compute TLSH hash for a character or raw vector and return hash fingerprint}
\usage{
tlsh_simple_hash(x)
}
\arguments{
\item{x}{length 1 \code{character} or \code{raw} vector}
}
\description{
A warning will be issued if the input byte stream is <50 bytes.
}
